# server/api/openapi.yml
openapi: "3.0.3"
info:
  title: CV Builder API
  version: "1.0"
security:
  - cookieAuth: []
paths:
...
  /login:
    post:
      operationId: PostAuthLogin
      security: []
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /register:
    post:
      operationId: PostAuthRegister
      security: []
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        "200":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthUser:
      type: object
      required: [id, email, plan]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        plan:
          type: string
          enum: [free, mid, pro]
...